fixing type imports in case domain as I moved the type definitions to the store
